+++ src/main/java/com/ettiv/docs/domain/DocRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface DocRepository extends PagingAndSortingRepository<Doc, Integer> {
+public interface DocRepository extends PagingAndSortingRepository<Doc, Long> {
 
     List<Doc> findByName(@Param("name") String name);
