+++ src/main/java/com/ettiv/docs/domain/EmployeeRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface EmployeeRepository extends PagingAndSortingRepository<Employee, Integer> {
+public interface EmployeeRepository extends PagingAndSortingRepository<Employee, Long> {
 
     List<Employee> findByFio(@Param("fio") String fio);
