+++ src/main/java/com/ettiv/docs/domain/OrgRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface OrgRepository extends PagingAndSortingRepository<Org, Integer> {
+public interface OrgRepository extends PagingAndSortingRepository<Org, Long> {
 
     List<Org> findByName(@Param("name") String name);
