+++ src/main/java/com/ettiv/docs/domain/UnitRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface UnitRepository extends PagingAndSortingRepository<Unit, Integer> {
+public interface UnitRepository extends PagingAndSortingRepository<Unit, Long> {
 
     List<Unit> findByName(@Param("name") String name);
