+++ src/main/java/com/ettiv/docs/domain/VidDocRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(collectionResourceRel = "vids", path = "vids")
-public interface VidDocRepository extends PagingAndSortingRepository<VidDoc, Integer> {
+public interface VidDocRepository extends PagingAndSortingRepository<VidDoc, Long> {
 
     List<VidDoc> findByName(@Param("name") String name);
